+++ test/standalone/Vegetation/test_bigleaf_parameterizations.jl
@@ -151,10 +151,10 @@ for FT in (Float32, Float64)
                 Γstar,
             )
         @test all(@.(Aj == J * (ci - Γstar) / (4 * (ci + 2 * Γstar))))
-        β = moisture_stress(
+        β = compute_tuzet_moisture_stress(
             p_l,
-            photosynthesisparams.sc,
             photosynthesisparams.pc,
+            photosynthesisparams.sc,
         )
         @test β ==
               (1 + exp(photosynthesisparams.sc * photosynthesisparams.pc)) / (